Split runtime integration tests and extract shared harness.
Centralize the runtime test harness and scenario builders under test/support, and replace the 3.6k-line index.test.ts with focused behavior suites so failures are easier to localize.

Closes #24
